requirements.txt
Pillow~=9.1.1
setuptools~=60.2.0
